requirements.txt
django==3.*
django-debug-toolbar
django-extensions
ipython
django-admin-rangefilter
django-import-export
django-storages
boto3
Celery
requests
django-filter
djangorestframework
djangorestframework-yaml
djangorestframework-xml
djangorestframework-simplejwt
drf-yasg

# flake8
flake8
flake8-builtins
flake8-import-order
flake8-print
flake8-breakpoint

# pytest
pytest
pytest-django
pytest-cov
pytest-mock